Support non-BB LLVM build for Emscripten
When targeting Emscripten/wasm, we were in the odd position that
we had BinaryBuilder support, but not in-tree support for building
LLVM. Unfortunately, the BinaryBuilder version has recently stopped
working due to ABI changes in Emscripten. Use that as an opportunity
to backfill the requisite definitions to make an in-tree build of
LLVM work on Emscripten.